% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{calculate_f1_score}
\alias{calculate_f1_score}
\title{Calculate F1 Score}
\usage{
calculate_f1_score(predictions, actual)
}
\arguments{
\item{predictions}{A numeric vector of predicted values, where 1 indicates a positive prediction and 0 indicates a negative prediction.}

\item{actual}{A numeric vector of actual outcomes, where 1 indicates a positive outcome and 0 indicates a negative outcome.}
}
\value{
A numeric value representing the F1 score.
}
\description{
Computes the F1 score, which is the harmonic mean of precision and recall, for a given set of predictions and actual outcomes.
}
\details{
The F1 score is calculated using the formula:
\deqn{F1 = \frac{2 \times \text{precision} \times \text{recall}}{\text{precision} + \text{recall}}}
where precision is the ratio of true positives to the total number of predicted positives, and recall is the ratio of true positives to the actual number of positives.
}
\examples{
# Example usage
predictions <- c(1, 0, 1, 1, 0)
actual <- c(1, 0, 0, 1, 1)
calculate_f1_score(predictions, actual)

}
