% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{calculate_swap_probability}
\alias{calculate_swap_probability}
\title{Calculate Swap Probability}
\usage{
calculate_swap_probability(score1, score2, temp1, temp2)
}
\arguments{
\item{score1}{Numeric. The score of the first state.}

\item{score2}{Numeric. The score of the second state.}

\item{temp1}{Numeric. The temperature associated with the first state.}

\item{temp2}{Numeric. The temperature associated with the second state.}
}
\value{
A numeric value between 0 and 1 representing the probability of swapping the two states.
The probability is calculated using the formula \eqn{\exp((score2 - score1) / (temp_high - temp_low))},
where \eqn{temp_high} and \eqn{temp_low} are the higher and lower temperatures of the two states, respectively.
}
\description{
This function calculates the probability of swapping two states based on their scores and temperatures.
It uses the Metropolis criterion to determine the swap probability.
}
\examples{
# Example usage:
calculate_swap_probability(score1 = 0.8, score2 = 0.9, temp1 = 0.5, temp2 = 0.6)

}
