% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{computeScores}
\alias{computeScores}
\title{Compute Scores with Logic}
\usage{
computeScores(data, columns, outcome, previous_rule, previous_rule_name)
}
\arguments{
\item{data}{A data frame containing the features and the outcome variable.}

\item{columns}{A character vector specifying the names of the columns to be used in the computation.}

\item{outcome}{A character string specifying the name of the outcome variable.}

\item{previous_rule}{A logical vector representing the previous rule applied. Default is NULL.}

\item{previous_rule_name}{A character string specifying the name of the previous rule. Default is an empty string.}
}
\value{
A list containing the computed scores and the corresponding rules.
}
\description{
This function computes scores for given data and columns using different methods
based on the type of outcome (binary or continuous).
}
\examples{
scores <- computeScores(data = df, columns = colnames(X), outcome = "outcome", previous_rule = NULL, previous_rule_name = "")
}
