% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{count_logical_operators}
\alias{count_logical_operators}
\title{Count Logical Operators}
\usage{
count_logical_operators(rule)
}
\arguments{
\item{rule}{A character string representing the rule.}
}
\value{
An integer representing the total count of logical operators in the rule.
}
\description{
This function counts the number of logical operators (& and |) in a given rule string.
}
\examples{
num_operators <- count_logical_operators("(var1 == 1) & (var2 == 1) | (var3 == 1)")
}
