% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{evaluate_all_rules}
\alias{evaluate_all_rules}
\title{Evaluate a List of Rules on Data}
\usage{
evaluate_all_rules(rules, data)
}
\arguments{
\item{rules}{A character vector containing the logical rules to be evaluated.}

\item{data}{A data frame containing the data on which the rules will be evaluated.}
}
\value{
A data frame where each column corresponds to the evaluated result (0/1) of each rule on the data.
}
\description{
This function evaluates a list of logical rules on a given dataset and stores the results in a DataFrame.
}
\examples{
rules <- c("(KRFP1536 | KRFP4081)", "((KRFP1536 == 1 | KRFP4081 == 1) & (KRFP3752 | Not_KRFP138))")
data <- data.frame(KRFP1536 = c(0, 1), KRFP4081 = c(1, 0), KRFP3752 = c(0, 1), Not_KRFP138 = c(1, 0))
evaluate_all_rules(rules, data)
}
