% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlis.R
\name{fit_lasso}
\alias{fit_lasso}
\title{Fit a Lasso Regression Model}
\usage{
fit_lasso(features, outcome, family)
}
\arguments{
\item{features}{A matrix of predictor variables. Each column represents a predictor, and each row represents an observation.}

\item{outcome}{A vector of response variables. The length of \code{outcome} should match the number of rows in \code{features}.}

\item{family}{A character string specifying the error distribution and link function to be used in the model. Options include \code{"gaussian"}, \code{"binomial"}, \code{"poisson"}, etc.}
}
\value{
A \code{cv.glmnet} model object containing the fitted Lasso regression model.
}
\description{
This function fits a Lasso regression model using cross-validated \code{glmnet}.
}
\examples{
\dontrun{
# Example with gaussian family
features <- matrix(rnorm(100*10), 100, 10)
outcome <- rnorm(100)
lasso_model <- fit_lasso(features, outcome, family = "gaussian")

# Example with binomial family
features <- matrix(rnorm(100*10), 100, 10)
outcome <- sample(c(0, 1), 100, replace = TRUE)
lasso_model <- fit_lasso(features, outcome, family = "binomial")
}

}
