% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_iteration.R
\name{run_single_iteration}
\alias{run_single_iteration}
\title{Run a Single Iteration of the Logic Tree Building Process}
\usage{
run_single_iteration(
  data,
  outcome,
  columns,
  max_operators,
  max_trees,
  temperatures,
  two_way_logic_roots,
  beam_width,
  no_improvement_threshold
)
}
\arguments{
\item{data}{Data frame. The dataset to be used for building the logic tree.}

\item{outcome}{Character. The name of the outcome variable.}

\item{columns}{Character vector. The names of the columns to be used for building the logic rules.}

\item{max_operators}{Integer. The maximum depth of the logic tree.}

\item{initial_temperature}{Numeric. The initial temperature for the simulated annealing process.}

\item{min_temperature}{Numeric. The minimum temperature in the middle of the iterations.}

\item{max_iterations}{Integer. The maximum number of iterations to run the logic tree building process.}
}
\value{
List. A list containing the best scores and best rule paths found during the iteration.
}
\description{
This function runs a single iteration of the logic tree building process, which involves recursively
building logic trees with a U-shaped temperature schedule for simulated annealing.
}
\examples{
data <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = sample(0:1, 100, replace = TRUE))
run_single_iteration(data, outcome = "y", columns = c("x1", "x2"), max_operators = 3, initial_temperature = 1, min_temperature = 0.01, max_iterations = 100)

}
